+++ CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 0.8.0 - 08/03/2022
+
 ### Added
 
  - [JAP-1](https://makeitapp.atlassian.net/browse/JAP-2): Added `/grant/bindings/resource/{resourceType}` api to handle the grant of a user role on a single resource
-  
  - [JAP-2](https://makeitapp.atlassian.net/browse/JAP-2): Added `/revoke/bindings/resource/{resourceType}` api to handle the revoke of a user role on a set resources
 
 ## 0.7.0 - 02/03/2022